feat: implement EVM validator registration and resignation
