--- src/domains/transaction/pages/SendDelegateResignation/index.ts
@@ -1,3 +1,0 @@
-export * from "./FormStep";
-export * from "./ReviewStep";
-export { SendDelegateResignation as default } from "./SendDelegateResignation";
