+++ src/domains/transaction/pages/SendValidatorResignation/ReviewStep.tsx
@@ -1,28 +1,35 @@
-import { Contracts as ProfilesContracts } from "@ardenthq/sdk-profiles";
-import React from "react";
+import { Contracts } from "@ardenthq/sdk-profiles";
+import React, { useEffect } from "react";
+import { useFormContext } from "react-hook-form";
 import { useTranslation } from "react-i18next";
 
-import { Alert } from "@/app/components/Alert";
-import { FormField, FormLabel } from "@/app/components/Form";
-import { FeeField } from "@/domains/transaction/components/FeeField";
+import { TotalAmountBox } from "@/domains/transaction/components/TotalAmountBox";
 import { TransactionAddresses } from "@/domains/transaction/components/TransactionDetail";
 import { StepHeader } from "@/app/components/StepHeader";
+import { ThemeIcon } from "@/app/components/Icon";
 import { DetailTitle, DetailWrapper } from "@/app/components/DetailWrapper";
 import { Divider } from "@/app/components/Divider";
-import { ThemeIcon } from "@/app/components/Icon";
 
-interface FormStepProperties {
-	senderWallet: ProfilesContracts.IReadWriteWallet;
-	profile: ProfilesContracts.IProfile;
-}
-
-export const FormStep = ({ senderWallet, profile }: FormStepProperties) => {
+export const ReviewStep = ({
+	senderWallet,
+	profile,
+}: {
+	senderWallet: Contracts.IReadWriteWallet;
+	profile: Contracts.IProfile;
+}) => {
 	const { t } = useTranslation();
 
+	const { getValues, unregister } = useFormContext();
+
+	useEffect(() => {
+		unregister("mnemonic");
+	}, [unregister]);
+
 	return (
-		<section data-testid="SendDelegateResignation__form-step" className="space-y-6 sm:space-y-4">
+		<section data-testid="SendValidatorResignation__review-step">
 			<StepHeader
-				title={t("TRANSACTION.PAGE_VALIDATOR_RESIGNATION.FORM_STEP.TITLE")}
+				title={t("TRANSACTION.REVIEW_STEP.TITLE")}
+				subtitle={t("TRANSACTION.REVIEW_STEP.DESCRIPTION")}
 				titleIcon={
 					<ThemeIcon
 						dimensions={[24, 24]}
@@ -32,21 +39,16 @@ export const FormStep = ({ senderWallet, profile }: FormStepProperties) => {
 						greenLightIcon="SendTransactionLightGreen"
 					/>
 				}
-				subtitle={t("TRANSACTION.PAGE_VALIDATOR_RESIGNATION.FORM_STEP.DESCRIPTION")}
 			/>
 
-			<Alert>{t("TRANSACTION.PAGE_VALIDATOR_RESIGNATION.FORM_STEP.WARNING")}</Alert>
-
-			<div className="space-y-3 sm:space-y-4">
-				<FormField name="senderAddress">
-					<TransactionAddresses
-						senderAddress={senderWallet.address()}
-						network={senderWallet.network()}
-						recipients={[]}
-						profile={profile}
-						labelClassName="w-auto sm:min-w-32"
-					/>
-				</FormField>
+			<div className="-mx-3 space-y-3 sm:mx-0 sm:space-y-4">
+				<TransactionAddresses
+					labelClassName="w-auto sm:min-w-28"
+					senderAddress={senderWallet.address()}
+					network={senderWallet.network()}
+					recipients={[]}
+					profile={profile}
+				/>
 
 				<DetailWrapper label={t("TRANSACTION.TRANSACTION_TYPE")}>
 					<div className="space-y-3 sm:space-y-0">
@@ -72,15 +74,9 @@ export const FormStep = ({ senderWallet, profile }: FormStepProperties) => {
 					</div>
 				</DetailWrapper>
 
-				<FormField name="fee">
-					<FormLabel>{t("TRANSACTION.TRANSACTION_FEE")}</FormLabel>
-					<FeeField
-						type="delegateResignation"
-						data={undefined}
-						network={senderWallet.network()}
-						profile={profile}
-					/>
-				</FormField>
+				<div className="mx-3 mt-2 sm:mx-0">
+					<TotalAmountBox amount={0} fee={getValues("fee")} ticker={senderWallet.currency()} />
+				</div>
 			</div>
 		</section>
 	);
