+++ src/domains/transaction/pages/SendValidatorResignation/index.ts
@@ -1,0 +1,3 @@
+export * from "./FormStep";
+export * from "./ReviewStep";
+export { SendValidatorResignation as default } from "./SendValidatorResignation";
