+++ src/domains/transaction/routing.ts
@@ -3,7 +3,7 @@ import { ProfilePaths } from "@/router/paths";
 import { RouteItem } from "@/router/router.types";
 import preloadLazy from "@/utils/preload-lazy";
 
-const SendDelegateResignation = preloadLazy(() => import("./pages/SendDelegateResignation"));
+const SendValidatorResignation = preloadLazy(() => import("./pages/SendValidatorResignation"));
 const SendRegistration = preloadLazy(() => import("./pages/SendRegistration"));
 const SendTransfer = preloadLazy(() => import("./pages/SendTransfer") as Promise<{ default: ComponentType<unknown> }>);
 const SendVote = preloadLazy(() => import("./pages/SendVote"));
@@ -15,9 +15,9 @@ export const TransactionRoutes: RouteItem[] = [
 		path: ProfilePaths.SendRegistration,
 	},
 	{
-		component: SendDelegateResignation,
+		component: SendValidatorResignation,
 		exact: true,
-		path: ProfilePaths.SendDelegateResignation,
+		path: ProfilePaths.SendValidatorResignation,
 	},
 	{
 		component: SendTransfer,
